# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Health & Nutrition Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /daily-logs:
    get:
      operationId: getDailyLogs
      tags: [daily-logs]
      summary: List daily logs
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
            default: 30
      responses:
        "200":
          description: List of daily logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyLog"
    post:
      operationId: createDailyLog
      tags: [daily-logs]
      summary: Create or upsert a daily log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
  /daily-logs/{date}:
    get:
      operationId: getDailyLog
      tags: [daily-logs]
      summary: Get daily log by date (YYYY-MM-DD)
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
        "404":
          description: Not found
    patch:
      operationId: updateDailyLog
      tags: [daily-logs]
      summary: Update a daily log by date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyLogUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
    delete:
      operationId: deleteDailyLog
      tags: [daily-logs]
      summary: Delete a daily log
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /meals/analyze-photo:
    post:
      operationId: analyzeMealPhoto
      tags: [meals]
      summary: Analyze a meal photo with AI and return suggested food items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPhotoAnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPhotoAnalysis"
...
  /advisor/check:
    post:
      operationId: advisorCheck
      tags: [advisor]
      summary: Rules-based meal guidance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvisorQuery"
      responses:
        "200":
          description: Advisor response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvisorResponse"
...
components:
  schemas:
...
    DailyLog:
      type: object
      required: [id, date]
      properties:
        id:
          type: number
        date:
          type: string
        calories:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
        water:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        steps:
          type: ["number", "null"]
        workoutMinutes:
          type: ["number", "null"]
        workoutType:
          type: ["string", "null"]
        sleepHours:
          type: ["number", "null"]
        reflux:
          type: ["number", "null"]
        postMealSleepiness:
          type: ["number", "null"]
        energy:
          type: ["number", "null"]
        headache:
          type: ["number", "null"]
        stress:
          type: ["number", "null"]
        bowelMovement:
          type: ["string", "null"]
        hungerBeforeLunch:
          type: ["number", "null"]
        hungerBeforeDinner:
          type: ["number", "null"]
        hungerBeforeBed:
          type: ["number", "null"]
        muscleStiffness:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DailyLogInput:
      type: object
      required: [date]
      properties:
        date:
          type: string
        calories:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
        water:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        steps:
          type: ["number", "null"]
        workoutMinutes:
          type: ["number", "null"]
        workoutType:
          type: ["string", "null"]
        sleepHours:
          type: ["number", "null"]
        reflux:
          type: ["number", "null"]
        postMealSleepiness:
          type: ["number", "null"]
        energy:
          type: ["number", "null"]
        headache:
          type: ["number", "null"]
        stress:
          type: ["number", "null"]
        bowelMovement:
          type: ["string", "null"]
        hungerBeforeLunch:
          type: ["number", "null"]
        hungerBeforeDinner:
          type: ["number", "null"]
        hungerBeforeBed:
          type: ["number", "null"]
        muscleStiffness:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    DailyLogUpdate:
      type: object
      properties:
        calories:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
        water:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        steps:
          type: ["number", "null"]
        workoutMinutes:
          type: ["number", "null"]
        workoutType:
          type: ["string", "null"]
        sleepHours:
          type: ["number", "null"]
        reflux:
          type: ["number", "null"]
        postMealSleepiness:
          type: ["number", "null"]
        energy:
          type: ["number", "null"]
        headache:
          type: ["number", "null"]
        stress:
          type: ["number", "null"]
        bowelMovement:
          type: ["string", "null"]
        hungerBeforeLunch:
          type: ["number", "null"]
        hungerBeforeDinner:
          type: ["number", "null"]
        hungerBeforeBed:
          type: ["number", "null"]
        muscleStiffness:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...
    MealPhotoAnalysisInput:
      type: object
      required: [imageBase64]
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image data (no data URL prefix)
        mediaType:
          type: ["string", "null"]
          description: Image MIME type, e.g. image/jpeg
    MealPhotoAnalysis:
      type: object
      required: [dishName, foodItems]
      properties:
        dishName:
          type: string
        mealType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalCalories:
          type: ["number", "null"]
        totalProtein:
          type: ["number", "null"]
        totalCarbs:
          type: ["number", "null"]
        totalFat:
          type: ["number", "null"]
        totalFiber:
          type: ["number", "null"]
        foodItems:
          type: array
          items:
            $ref: "#/components/schemas/AnalyzedFoodItem"
    AnalyzedFoodItem:
      type: object
      required: [name]
      properties:
        name:
          type: string
        portion:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        calories:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
...
    AdvisorQuery:
      type: object
      required: [question, date]
      properties:
        question:
          type: string
        date:
          type: string
        foodName:
          type: ["string", "null"]
        portionDescription:
          type: ["string", "null"]
        estimatedCalories:
          type: ["number", "null"]
        estimatedProtein:
          type: ["number", "null"]
        estimatedCarbs:
          type: ["number", "null"]
        estimatedFiber:
          type: ["number", "null"]
    AdvisorResponse:
      type: object
      required: [decision, explanation, tips]
      properties:
        decision:
          type: string
        explanation:
          type: string
        remainingCalories:
          type: ["number", "null"]
        remainingProtein:
          type: ["number", "null"]
        remainingCarbs:
          type: ["number", "null"]
        remainingFiber:
          type: ["number", "null"]
        remainingWater:
          type: ["number", "null"]
        tips:
          type: array
          items:
            type: string
...